+++ lib/minitest/event_prof_formatter.rb
@@ -26,11 +26,13 @@ module Minitest
       private
 
       def total_results(profiler)
+        time_percentage = time_percentage(profiler.total_time, profiler.absolute_run_time)
+
         @results <<
           <<~MSG
             EventProf results for #{profiler.event}
 
-            Total time: #{profiler.total_time.duration}
+            Total time: #{profiler.total_time.duration} of #{profiler.absolute_run_time.duration} (#{time_percentage}%)
             Total events: #{profiler.total_count}
 
             Top #{profiler.top_count} slowest suites (by #{profiler.rank_by}):
@@ -39,30 +41,46 @@ module Minitest
       end
 
       def by_groups(profiler)
-        profiler.results[:groups].each do |group|
+        result = profiler.results
+        groups = result[:groups]
+
+        groups.each do |group|
           description = group[:id][:name]
           location = group[:id][:location]
+          time = group[:time]
+          run_time = group[:run_time]
+          time_percentage = time_percentage(time, run_time)
 
           @results <<
             <<~GROUP
-              #{description.truncate} (#{location}) – #{group[:time].duration} (#{group[:count]} / #{group[:examples]})
+              #{description.truncate} (#{location}) – #{time.duration} (#{group[:count]} / #{group[:examples]}) of #{run_time.duration} (#{time_percentage}%)
             GROUP
         end
       end
 
       def by_examples(profiler)
-        return unless profiler.results[:examples]
+        result = profiler.results
+        examples = result[:examples]
+
+        return unless examples
         @results << "\nTop #{profiler.top_count} slowest tests (by #{profiler.rank_by}):\n\n"
 
-        profiler.results[:examples].each do |example|
+        examples.each do |example|
           description = example[:id][:name]
           location = example[:id][:location]
+          time = example[:time]
+          run_time = example[:run_time]
+          time_percentage = time_percentage(time, run_time)
 
           @results <<
             <<~GROUP
-              #{description.truncate} (#{location}) – #{example[:time].duration} (#{example[:count]})
+              #{description.truncate} (#{location}) – #{time.duration} (#{example[:count]}) of #{run_time.duration} (#{time_percentage}%)
             GROUP
         end
+      end
+
+      def time_percentage(time, total_time)
+        (time / total_time * 100).round(2)
       end
     end
   end