+++ CHANGELOG.md
@@ -10,6 +10,9 @@
  SAMPLE=10 rake test # runs 10 random test examples
  SAMPLE_GROUPS=10 rake test # runs 10 random example groups
  ```
+
+- Extend Event Prof formatter results that includes the absolute run time and percentage of event time percentage [#100](https://github.com/palkan/test-prof/issues/100) ([@dmagro][])
+
 ## 0.7.2 (2018-10-08)
 
 - Add `RSpec/AggregateFailures` support for non-regular 'its' examples. ([@broels][])
@@ -378,3 +381,4 @@ Ensure output dir exists in `#artifact_path` method.
 [@Vasfed]: https://github.com/Vasfed
 [@szemek]: https://github.com/szemek
 [@mkldon]: https://github.com/mkldon
+[@dmagro]: https://github.com/dmagro